Introduced intermediary interface CollectionValidator
Extracted methods for checking all the entries of a collection to
another interface. This way it is possible to handle Maps with most of
the available methods as well.

Map has key/value and not a single typed entry like Set and List so
checking all its content differs from List and Set.