Completely reordered class locations
The new order is this: class Validator (which is the entry point for
users) and all the interfaces for specialized Validators are located in
the base package.

ValidationException still has its own package

All implementations of Validators are located in package impl. They
shall not be visible by anyone outside of this package. That's why we
need the "entry-point-class" ValidatorImpl in impl so that class
Validator can access the implementations.