Improved unit test checks
During many of the unit tests the message of the ValidationException is
checked. Up to now it was only checked that a number of substrings is
contained in the message.

From now on it is also checked that these substrings appear in the given
order (because a sentence might mean something different when the words
are reordered).