removed specializations for int
Shame on me. Seems like I was in a programming spree and wrote methods
for (e.g.) lessThan(long) *and* lessThan(int).
To make it complete I also added tests for both methods m-)

This of course is completely useless and I removed these methods as well
as the extra tests.